make loading of tileset meta-data more robust
- check if the load is failed (MainWindow::openFile)
 - report error only if the file can not be read or the data is invalid
 - do not change the entity till the success is guaranteed
 - allow empty content
 - guess width/height of min-data only if there is data
 - add TILE_SIZE to use it instead of constants
 + fix createTile/createSubtile

+++ source/d1min.h
@@ -14,7 +14,7 @@ public:
     D1Min() = default;
     ~D1Min() = default;
 
-    bool load(QString minFilePath, quint16 subtileCount, std::map<unsigned, D1CEL_FRAME_TYPE> &celFrameTypes, const OpenAsParam &params);
+    bool load(QString minFilePath, int subtileCount, std::map<unsigned, D1CEL_FRAME_TYPE> &celFrameTypes, const OpenAsParam &params);
     bool save(D1Gfx *gfx, const SaveAsParam &params);
 
     QImage getSubtileImage(quint16);
